added more doctests
